+++ R/get_site_general.R
@@ -1,5 +1,6 @@
 #' @title eLTER get_site_general function
-#' @description This function obtains general information
+#' @description `r lifecycle::badge("stable")`
+#' This function obtains general information
 #' about an eLTER site through the DEIMS-SDR sites API.
 #' @param deimsid A `character`. The DEIMS ID of the site from
 #' DEIMS-SDR website. DEIMS ID information
@@ -16,20 +17,12 @@
 #'
 ### function get_site_general
 get_site_general <- function(deimsid) {
-  q <- '{title: .title,
-       uri: "\\(.id.prefix)\\(.id.suffix)",
-       geoCoord: .attributes.geographic.coordinates,
-       country: .attributes.geographic.country,
-       geoElev: .attributes.geographic.elevation,
-       generalInfo: .attributes.general
-      }'
-  jj <- get_id(deimsid, "sites")
+  qo <- queries_jq[[get_deims_API_version()]]$site_general
+  jj <- get_id(deimsid, qo$path)
   if (is.na(attr(jj, "status"))) {
     invisible(
       utils::capture.output(
-        general <- dplyr::as_tibble(
-          do_Q(q, jj)
-        )
+        general <- dplyr::as_tibble(do_Q(qo$query, jj))
       )
     )
     colnames(general$generalInfo.keywords[[1]]) <- c(