+++ tests/testthat/test-get_network_observedProperties.R
@@ -1,4 +1,4 @@
-message("\n---- Test get_network_parameters() ----")
+message("\n---- Test get_network_observedProperties() ----")
 
 library(testthat)
 
@@ -8,7 +8,7 @@ test_that("Expect error if internet connection is down", {
   Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   testthat::expect_error(
     httptest::without_internet(
-      result <- ReLTER::get_network_parameters(
+      result <- ReLTER::get_network_observedProperties(
         networkDEIMSID = TESTURLNetwork
       )
     ),
@@ -20,9 +20,9 @@ test_that("Expect error if internet connection is down", {
 skip_if_offline(host = "deims.org")
 skip_if(skip_in_test_mode)
 
-test_that("Output of network parameters function constructs 'tibble' as
+test_that("Output of network observed properties function constructs 'tibble' as
           expected", {
-  result <- ReLTER::get_network_parameters(
+  result <- ReLTER::get_network_observedProperties(
     networkDEIMSID = TESTURLNetwork
   )
   expect_s3_class(result, "tbl_df")
@@ -36,7 +36,7 @@ test_that("Output of network parameters function constructs 'tibble' as
 
 test_that("Wrong input (but URL) constructs a NULL object", {
   Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
-  result <- ReLTER::get_network_parameters(
+  result <- ReLTER::get_network_observedProperties(
     networkDEIMSID = "https://deims.org/network/ljhnhbkihubib"
   )
   expect_true(is.null(result))
@@ -47,7 +47,9 @@ test_that("Wrong input (but URL) constructs a NULL object", {
 
 test_that("Wrong input (not URL) constructs an empty tibble", {
   Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
-  result <- ReLTER::get_network_parameters(networkDEIMSID = "ljhnhbkihubib")
+  result <- ReLTER::get_network_observedProperties(
+    networkDEIMSID = "ljhnhbkihubib"
+  )
   expect_true(is.null(result))
   expect_true(is.null(ncol(result)))
   expect_true(length(result) == 0)