+++ NEWS.md
@@ -1,5 +1,23 @@
 # ReLTER 1.2.0 Release Notes
 
+<<<<<<< HEAD
+v1.2.0 was released on 01/07/2022
+
+## Changes
+
+* inserted new functions for set an environment for variables of the package,
+  for setting and for get the DEIMS-SDR base URL;
+* inserted a new function for get version of DEIMS-SDR API based on the
+  DEIMS-SDR base URL;
+* added an object (queries_jq) containing all the JQ queries (as a list);
+* added the new JQ queries accordingly with the structure of the new DEIMS-SDR
+  API version;
+* deleted the JQ queries call into the get_site_info(), get_activity_info(),
+  and get_dataset_info() and substitute with a variable;
+* changed the name of the functions containing "parameters" with "observed
+  properties";
+* added for some functions the life cycle badge.
+=======
 v1.1.0 was released on 15/09/2022
 
 ## Major changes
@@ -14,6 +32,7 @@ The new implemented functionalities concern:
 * interaction with Sensor Observations Services (SOS - v2.0) for acquire procedure list, procedure info, feature of interest info, observed property info, and observations;
 * interaction to the Zenodo repository in order to upload record or download dataset record.
 
+>>>>>>> dev
 
 ________________________________________________________________________________
 
@@ -31,7 +50,6 @@ v1.1.0 was released on 15/04/2022
 * enriched the information downloaded for datasets and activities from DEIMS-SDR;
 * a solution for install package by Docker has been done;
 * all the packages used in the `ReLTER` have been referenced in each functions manual.
-
 
 ________________________________________________________________________________